Initial commit, add a new global and some notes
As a starting point, I just shamelessly dumped in the
angular-cli.vim plugin and :%s to vionic.

I've been keeping notes in there for a while as well as
adding my own hacks (none of which I feel are worth a PR).

Leaving this here as a reminder to credit bdauria in the README